+++ Swing/app/src/main/java/com/kidsdynamic/swing/view/calendar/ViewCalendarCellWeek.java
@@ -55,7 +55,7 @@ public class ViewCalendarCellWeek extends ViewCalendarCell {
         delEvent(event.mId);
     }
 
-    public void delEvent(int id) {
+    public void delEvent(long id) {
         while (true) {
             int idx, count = mEventList.size();
             for (idx = 0; idx < count; idx++) {
